5.1-PP-data/pyproject.toml
[project]
name = "5-1-pp-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.11.3",
    "httpx>=0.28.1",
    "ipykernel>=6.30.1",
    "mcp[cli]>=1.13.1",
    "openai>=1.102.0",
    "python-dotenv>=1.1.1",
]
